Upgrade all dependencies, restructure project, bump to 1.4.0.
