+++ core/src/Test/resources/log4j.properties
@@ -1,0 +1,4 @@
+log4j.rootLogger=ERROR, CONSOLE
+log4j.appender.CONSOLE=org.apache.log4j.ConsoleAppender
+log4j.appender.CONSOLE.layout=org.apache.log4j.PatternLayout
+log4j.appender.CONSOLE.layout.ConversionPattern=[%-5p] %m%n
