+++ core/src/Test/scala/com/intenthq/pucket/writer/IncrementalPartitionedWriterSpec.scala
@@ -1,6 +1,6 @@
 package com.intenthq.pucket.writer
 
-import com.intenthq.pucket.{TestLogging, PucketDescriptor}
+import com.intenthq.pucket.PucketDescriptor
 import com.intenthq.pucket.TestUtils.PucketWrapper
 import org.apache.hadoop.fs.Path
 import org.specs2.Specification
@@ -9,7 +9,7 @@ import org.specs2.matcher.DisjunctionMatchers
 
 import scalaz.\/
 
-trait IncrementalPartitionedWriterSpec[T] extends Specification with DisjunctionMatchers with TestLogging {
+trait IncrementalPartitionedWriterSpec[T] extends Specification with DisjunctionMatchers {
   import IncrementalPartitionedWriterSpec._
 
   val wrapper: PucketWrapper[T]
