+++ core/src/Test/scala/com/intenthq/pucket/writer/PartitionedWriterSpec.scala
@@ -1,13 +1,13 @@
 package com.intenthq.pucket.writer
 
-import com.intenthq.pucket.{TestLogging, PucketDescriptor}
+import com.intenthq.pucket.PucketDescriptor
 import com.intenthq.pucket.TestUtils._
 import org.specs2.Specification
 import org.specs2.matcher.DisjunctionMatchers
 
 import scalaz.\/
 
-trait PartitionedWriterSpec[T] extends Specification with DisjunctionMatchers with TestLogging {
+trait PartitionedWriterSpec[T] extends Specification with DisjunctionMatchers {
   import PartitionedWriterSpec._
 
   val wrapper: PucketWrapper[T]
@@ -46,4 +46,3 @@ trait PartitionedWriterSpec[T] extends Specification with DisjunctionMatchers wi
 object PartitionedWriterSpec {
   val partitions = 0.to(2000)
 }
-
