+++ project/plugins.sbt
@@ -1,14 +1,8 @@
-resolvers ++= Seq(
-  "bigtoast-github" at "http://bigtoast.github.com/repo/",
-  "sbt-plugin-releases" at "http://repo.scala-sbt.org/scalasbt/sbt-plugin-releases"
-)
+resolvers += "bigtoast-github" at "http://bigtoast.github.com/repo/"
 
-addSbtPlugin("com.cavorite" % "sbt-avro" % "0.3.2")
-
+addSbtPlugin("com.cavorite" % "sbt-avro-1-7" % "1.1.2")
 addSbtPlugin("com.github.bigtoast" % "sbt-thrift" % "0.7")
-
-addSbtPlugin("com.jsuereth" % "sbt-pgp" % "1.0.0")
-
-addSbtPlugin("com.eed3si9n" % "sbt-unidoc" % "0.3.3")
-
-addSbtPlugin("com.timushev.sbt" % "sbt-updates" % "0.1.9")
+addSbtPlugin("com.jsuereth" % "sbt-pgp" % "1.0.1")
+addSbtPlugin("com.eed3si9n" % "sbt-unidoc" % "0.4.0")
+addSbtPlugin("com.timushev.sbt" % "sbt-updates" % "0.3.0")
+addSbtPlugin("net.virtual-void" % "sbt-dependency-graph" % "0.8.2")
